Revert ad-hoc mimicking of ANSI keyboard.
I have an ISO keyboard, not an ANSI keyboard. This means that the Enter
is L-shaped, and pretty far away. My muscle memory prefers ANSI, too. To
remedy this, I have attempted turning the bar/backslash key into an
additional Return, and hide the bar and backslash between the mode
switch for Return and forward slash, respectively. However, this
approach was uncomfortable, so I reverted it.